requirements.txt
appnope==0.1.0
click==6.6
decorator==4.0.10
defusedxml==0.4.1
Django==1.9.11
django-bootstrap-ui==0.2.0
django-localflavor==1.3
django-tag-parser==2.1
djangorestframework==3.4.6
dominate==2.1.17
huey==1.2.0
ipython==5.1.0
ipython-genutils==0.1.0
mysqlclient==1.3.7
oauthlib==1.1.2
ordereddict==1.1
pexpect==4.2.1
pickleshare==0.7.4
popuparchive==1.1.0
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
python-dotenv==0.5.1
python-social-auth==0.2.20
python3-openid==3.0.10
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
